# server/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Antimony API",
        "contact": {
            "name": "Institute for Networking at OST",
            "url": "https://www.ost.ch/en/research-and-consulting-services/computer-science/ins-institute-for-network-and-security",
            "email": "antimony@network.garden"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/topologies": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "topologies"
                ],
                "summary": "Get all topologies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.OkResponse-array_topology_TopologyOut"
                        }
                    },
                    "401": {
                        "description": "The user isn't authorized"
                    },
                    "403": {
                        "description": "Access to the resource was denied. Details in the request body.",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "498": {
                        "description": "The provided access token is not valid"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "topologies"
                ],
                "summary": "Create a new topology",
                "parameters": [
                    {
                        "description": "The topology",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/topology.TopologyIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The ID of the newly created collection",
                        "schema": {
                            "$ref": "#/definitions/utils.OkResponse-string"
                        }
                    },
                    "401": {
                        "description": "The user isn't authorized"
                    },
                    "403": {
                        "description": "Access to the resource was denied. Details in the request body.",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "498": {
                        "description": "The provided access token is not valid"
                    }
                }
            }
        },
        "/topologies/{id}": {
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "topologies"
                ],
                "summary": "Delete an existing topology",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The ID of the topology to delete",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "The user isn't authorized"
                    },
                    "403": {
                        "description": "Access to the resource was denied. Details in the request body.",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "The request was invalid. Details in the response body.",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "498": {
                        "description": "The provided access token is not valid"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "topologies"
                ],
                "summary": "Update an existing topology",
                "parameters": [
                    {
                        "description": "The topology with updated values",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/topology.TopologyIn"
                        }
                    },
                    {
                        "type": "string",
                        "description": "The ID of the topology to edit",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "The user isn't authorized"
                    },
                    "403": {
                        "description": "Access to the resource was denied. Details in the request body.",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "The request was invalid. Details in the response body.",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "498": {
                        "description": "The provided access token is not valid"
                    }
                }
            }
        },
        "/topologies/{topologyId}/files": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bindFiles"
                ],
                "summary": "Create a new bind file for a topology",
                "parameters": [
                    {
                        "description": "The bind file",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/topology.BindFileIn"
                        }
                    },
                    {
                        "type": "string",
                        "description": "The ID of the topology the bind file should belong to",
                        "name": "topologyId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The ID of the newly created file",
                        "schema": {
                            "$ref": "#/definitions/utils.OkResponse-string"
                        }
                    },
                    "401": {
                        "description": "The user isn't authorized"
                    },
                    "403": {
                        "description": "Access to the resource was denied. Details in the request body.",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "498": {
                        "description": "The provided access token is not valid"
                    }
                }
            }
        },
        "/topologies/{topologyId}/files/{bindFileId}": {
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bindFiles"
                ],
                "summary": "Update an existing bind file of a topology",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The ID of the topology the bind file belongs to",
                        "name": "topologyId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The ID of the bind file to edit",
                        "name": "bindFileId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "The user isn't authorized"
                    },
                    "403": {
                        "description": "Access to the resource was denied. Details in the request body.",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "The request was invalid. Details in the response body.",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "498": {
                        "description": "The provided access token is not valid"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bindFiles"
                ],
                "summary": "Delete an existing bind file of a topology",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The ID of the topology the bind file belongs to",
                        "name": "topologyId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The ID of the bind file to delete",
                        "name": "bindFileId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "The user isn't authorized"
                    },
                    "403": {
                        "description": "Access to the resource was denied. Details in the request body.",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "The request was invalid. Details in the response body.",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "498": {
                        "description": "The provided access token is not valid"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "topology.BindFileIn": {
            "type": "object",
            "required": [
                "filePath"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "filePath": {
                    "type": "string"
                }
            }
        },
        "topology.BindFileOut": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "filePath": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "topologyId": {
                    "type": "string"
                }
            }
        },
        "topology.TopologyIn": {
            "type": "object",
            "required": [
                "collectionId",
                "definition",
                "syncUrl"
            ],
            "properties": {
                "collectionId": {
                    "type": "string"
                },
                "definition": {
                    "type": "string"
                },
                "syncUrl": {
                    "type": "string"
                }
            }
        },
        "topology.TopologyOut": {
            "type": "object",
            "properties": {
                "bindFiles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/topology.BindFileOut"
                    }
                },
                "collectionId": {
                    "type": "string"
                },
                "creator": {
                    "$ref": "#/definitions/user.UserOut"
                },
                "definition": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastDeployFailed": {
                    "type": "boolean"
                },
                "syncUrl": {
                    "type": "string"
                }
            }
        },
...
        "user.UserOut": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "utils.OkResponse-array_topology_TopologyOut": {
            "type": "object",
            "properties": {
                "payload": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/topology.TopologyOut"
                    }
                }
            }
        },
...
        "utils.OkResponse-string": {
            "type": "object",
            "properties": {
                "payload": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}